+++ ui/src/shared/components/TimeMachine/TimeMachine.tsx
@@ -93,6 +93,7 @@ interface PassedProps {
   ) => JSX.Element
   queryStatus: QueryStatus
   onUpdateScriptStatus?: (status: ScriptStatus) => void
+  onActiveQueryIndexChange?: (activeQueryIndex: number) => void
   refresh: RefreshRate
 }
 
@@ -540,6 +541,10 @@ class TimeMachine extends PureComponent<Props, State> {
 
   private handleSetActiveQueryIndex = (activeQueryIndex): void => {
     this.setState({activeQueryIndex})
+    const {onActiveQueryIndexChange} = this.props
+    if (onActiveQueryIndexChange) {
+      onActiveQueryIndexChange(activeQueryIndex)
+    }
   }
 
   private handleSetActiveEditorTab = (tabName: CEOTabs): void => {